backend/requirements.txt
# Stock Technical Analysis AI - Requirements
# Updated: November 2025
# Flexible versions for better compatibility

# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data & Finance
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
#pandas-ta>=0.3.0
pandas-ta==0.4.71b0
ta>=0.10.0

# Chart Generation
mplfinance>=0.12.10b0
matplotlib>=3.7.0
plotly>=5.0.0
kaleido>=0.2.0

# AI & LLM (for Phase 2)
crewai>=0.11.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
anthropic>=0.7.0

# Document Generation
python-docx>=1.0.0
markdown>=3.5.0

# Utilities
requests>=2.31.0
Pillow>=10.0.0
aiohttp>=3.9.0
httpx>=0.25.0